requirements.txt
av==12.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.3.1
decorator==4.4.2
faster-whisper==1.0.2
ffmpeg-python==0.2.0
filelock==3.15.1
flatbuffers==24.3.25
fsspec==2024.6.0
future==1.0.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
mpmath==1.3.0
numpy==2.0.0
onnxruntime==1.18.0
packaging==24.1
pillow==10.3.0
proglog==0.1.10
protobuf==5.27.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
sympy==1.12.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
